Rename folders so it is clearer what they contain Hopefully, this will be the last time I change them.
The actual driver is a Virtual USB host controller that imports a USB
device from a USB over IP capable server. We could create a VUDC driver at
some point if we want to export USB devices.

The userspace is the usbip-utils for Windows, which should be pretty much
as the same set of tools that can be found under Linux.

Signed-off-by: Alexandre Demers <alexandre.f.demers@gmail.com>