Properly fix cancel_irp() routine
The previous commit was based on the assumption that we could not know at
which IRQL we were accessing cancel_irp(). This was false since Microsoft's
documentation states that IoCancelIrp() will call the cancel IRP routine
and that the I/O manager will make sure we are running at DISPATCH_LEVEL.

However, the real problem is more about how we handle the spinlocks and
what needs to be done by the cancel IRP routine. This commit should now
properly handle what is required according to Windows' driver
documentation. The spinlock must be released before calling
IoCompleteRequest(), which we were doing too late, otherwise we can end up
with a deadlock. Also, the cancel routine must set the IoStatus
Information member to 0, which was missing.

So we release the global cancel spinlock early, but stay at DISPATCH_LEVEL,
allowing the use of KeAcquireSpinLockAtDpcLevel on our own lock on queue.
However, we will release the lock at the IRQL that was set when the
I/O manager called the cancel_irp() routine after searching for the IRP to
be canceled in our list. Then, we finish cleaning up, before calling
IoCompleteRequest().

Signed-off-by: Alexandre Demers <alexandre.f.demers@gmail.com>