[README] Add some updates on the development
Yes, this project has been on hold for the last few months, but... some work has been done and is mostly ready to push: porting the userspace tools from Linux to Windows. Yeah!

Also some news on why I have been slow to work on this project and when the work should be resumed.